Updater PricingPage tests
